# Eliminate skipped steps due to inaccurate pulse timing