### Description

This is a rework of the way step pulses are timed, so that a minimum step can be guaranteed.

I am sure there will be some concerns and discussion surrounding this approach. While solving some problems, it introduces some complications, which I mention later in the PR. I will be away for a week, so wanted to post what I have so people can begin exploring the idea, or try it out on their printers if they find it promising.

I don't consider the code cleanup complete by any means (especially in stepper.h). I expect to need to put additional work into this when I get back, if we choose to use it.

#### How does it work before this change?

1. Timer is read at the beginning of the stepper loop.
2. Offsets of all step transitions are calculated from initial timer value using hard-coded cycle counts.
3. Interruptions such as systick of software serial can interfere, without changing the timer ticks at which steps transitions will occur.

**Example of pulse which is shorter than MINIMUM_STEPPER_PULSE (DRV8825)**
Configured 2us pulse is < 1us due to SysTick interference during timer access.
![image](https://user-images.githubusercontent.com/20053467/70313108-d3f2ee00-17c9-11ea-9bec-234fba4191a5.png)

**Example of short pulse (and missed step) during multi-stepping**
![image](https://user-images.githubusercontent.com/20053467/70313422-73b07c00-17ca-11ea-94d8-e1d8f2e3980b.png)


#### How does it work after this change?

1. Each stepper pulse edge is timed relative to the previous edge.
2. The only constant cycle counts needed are the times to work with the timer.
3. Some logic from PULSE_START was extracted to PULSE_PREP.
    - Reduces "spread" from the rising edge of the first stepper to the rising edge of the last, allowing less error when timing after the last rising edge.
    - PULSE_PREP work is performed during the low stage of the prior pulse.
    - Allows logic to be "hidden" inside pulses when multi-stepping.
    - Allows timer interactions to be closer to pulse edges.
    - XYZ "spread" on AVR is reduced from ~7.7us to ~700ns
    - XYZ "spread" on LPC1768 is reduced from ~300ns to ~79ns

XYZ Rise-Time Spread before change, on LPC1768:
![image](https://user-images.githubusercontent.com/20053467/70313970-6c3da280-17cb-11ea-977a-65a466fc2061.png)

XYZ Rise-Time Spread after change, on LPC1768:
![image](https://user-images.githubusercontent.com/20053467/70314013-88414400-17cb-11ea-8de3-869d7b4c6762.png)


### Benefits

- Guarantees minimum duration of step, as defined in MINIMUM_STEPPER_PULSE, resolving skipped steps due to inaccurate pulse timing.
- Reduces dependence of hard-to-maintain timing constants (also a concern)
- Improves pulse consistency, regardless of the location they come from (stepper ISR, Linear Advance, etc.)
- Allows sub-1us pulse periods by setting a maximum rate without a minimum period.

### Concerns

- There is currently a very small timing degradation, on some combinations of hardware. This can likely be solved with a little more tuning I didn't have time for before posting this.
- Cycle-count constants are still used to calculate maximum step rates. This change doesn't update those, and they will still be hard to maintain as anything changes inside the stepper ISR. I'm really not sure how to deal with this right now.
- TMC step rates are not as fast as I would like. I foresee a future change on top of this to optimize for very short pulse durations, such as the 100ns acceptable to TMC drivers. This would bypass timer usage, as querying the timer is quite slow on all platforms.

### Related Issues

#11205 - User tested an early experiment with a similar solution
#11047 - Probably will be fixed
#12403 - Fairly broad issue, likely resolves some of the issues, especially where sBase is mentioned