@sjasonsmith 2019-12-06
Here is a picture showing the behavior change with an actual print. These are the same gcode printed with the same printer configuration. The only difference is that the cube on the write was printed with firmware including this PR.

These were printed on an MKS sBase, which has built-in 8825 drivers. The printer is not very good or well-tuned, so there are some other artifacts in the print, but the main point is the straightness of the edges and consistency of extrusion.

![image](https://user-images.githubusercontent.com/20053467/70317877-854a5180-17d3-11ea-9dec-d6c1c5af5b97.png)