@ManuelMcLure 2019-12-06
How does this interact with ``SQUARE_WAVE_STEPPING``?