@sjasonsmith 2019-12-06
> How does this interact with `SQUARE_WAVE_STEPPING`?

This is a good question. I didn't explicitly test it (yet). Thanks for the reminder.

At one point I had it working so that specifying MINIMUM_STEPPER_PULSE=0 along with a MAXIMUM_STEPPER_RATE would essentially skip the first delay, but I think that behavior was lost when I reworked things to be more AVR-friendly. I will have to revisit that scenario.

There is a lot of room for TMC improvement, in general. The timer interactions are quite slow, and I can get higher performance by simply delaying 100ns, rather than reading a timer, which takes about 230ns on LPC1768. I went back to the timer approach to avoid exploding AVR, where it is extremely important to perform as much logic as possible while waiting for timers.

If this goes forward, I expect a follow-up to improve speed on 32-bit. It just added a lot of noise to this change trying to do everything at once.