@sjasonsmith 2019-12-07
I was susprised by the error on STM32 about F_CPU not being constant, but I see similar issues have been encountered before. That makes sense when considering the chip can probably be slowed down to save power.