@thinkyhead 2019-12-08
Thank you for taking a close look at this. Marlin's current approach is definitely imperfect, and I'm not surprised to see missed (micro) steps are happening.

In an ideal world MCUs would give you a pin mode where the output turns itself LOW after some duration whenever it gets turned HIGH (or vice-versa)…. Maybe in the future.

I'll review your approach more closely and look at the trade-offs. Pulse duration is a fundamental issue, so it will be good to get our algorithms working solidly.