@sjasonsmith 2019-12-08
I'll be down in Austin all of next week so won't be able to try out any new ideas right away. I should be able to get on here semi-regularly to discuss ideas and concerns.

I fully expect some controversy, as this is a significant paradigm shift to the timing mechanisms in this ISR, and I would be suspicious if there were _not_ concerns.

Right now my largest concern is the maximum stepper rate calculations. I would love to eliminate all those hard-coded cycle counts, and the maintenance that goes with them. Maintaining those cycle counts create a huge barrier of entry to anyone wanting to improve this code, and are likely not consistently perfect as compilers and processor architectures change.