@ejtagle 2019-12-08
> In an ideal world MCUs would give you a pin mode where the output turns itself LOW after some duration whenever it gets turned HIGH (or vice-versa)…. Maybe in the future.

Most MCUs have that exact ability , but just for selected pins....