@thinkyhead 2019-12-11
> I'll be down in Austin all of next week so won't be able to try out any new ideas right away. I should be able to get on here semi-regularly to discuss ideas and concerns.

I'm in Austin all week too. All the time, really….