@thinkyhead 2019-12-11
> Most MCUs have that exact ability , but just for selected pins....

I wish we only had to support a single MCU and board, already organized with all these factors in mind. But of course we have to maintain some time-checking method for AVR (where it's adequate) and then we need to provide some time-based method for any STEP pins that don't have this special mode. So it's another thing we could check for as we now do with PWM, then use whenever possible….

Looking to the future, modern boards will benefit from a 32-bit-only firmware built from the ground up to:

- allocate pins and their modes to features / functions, then
- abstract the needed protocols on top of that layer, and then
- add the needed abstract interfaces on top of that layer.

…where we have our own methods to allocate SW SPI and SW Serial as needed.

Two approaches rise to the top:
- Re-architect Marlin around an RTOS that has broad platform support
- Add a HAL to RepRapFirmware, then port over our hardware support and features.

So far my preference is to build on top of RepRapFirmware because it has already solved some low level problems and provides a framework for adding new features. However, it doesn't have a full Hardware Abstraction Layer at the level of Marlin, but it supports a couple of ATSAM boards.