@ejtagle 2019-12-13
> > Most MCUs have that exact ability , but just for selected pins....
> 
> I wish we only had to support a single MCU and board, already organized with all these factors in mind. But of course we have to maintain some time-checking method for AVR (where it's adequate) and then we need to provide some time-based method for any STEP pins that don't have this special mode. So it's another thing we could check for as we now do with PWM, then use whenever possible….
> 
> Looking to the future, modern boards will benefit from a 32-bit-only firmware built from the ground up to:
> 
> * allocate pins and their modes to features / functions, then
> * abstract the needed protocols on top of that layer, and then
> * add the needed abstract interfaces on top of that layer.
> 
> …where we have our own methods to allocate SW SPI and SW Serial as needed.
> 
> Two approaches rise to the top:
> 
> * Re-architect Marlin around an RTOS that has broad platform support
> * Add a HAL to RepRapFirmware, then port over our hardware support and features.
> 
> So far my preference is to build on top of RepRapFirmware because it has already solved some low level problems and provides a framework for adding new features. However, it doesn't have a full Hardware Abstraction Layer at the level of Marlin, but it supports a couple of ATSAM boards.

I do not know if it call my experience fortunate or infortunate, but right now, there is no vendor supplied HAL that targets all the MPUs that are being brought into the market. 

RTOS support for new MCUs is lagging quite a bit, so moving to an RTOS would not solve anything for newer MCUs, as we would have to add support for them ourselves.

So, i think your idea of building over Reprap firmware, it probably the most sane approach here. Marlin has to support a ton of different hardware platforms, thus, most of the time, there is simply no way to ensure specific hw features (such as PWM on STEP pins) is available.

So, most of the time a software fallback must exist (as Marlin already does with mostly everything)

The only thing you usually can count on are at least having 1 o 2 timers to create periodic interrupts... This pull request tries to improve the pulse width without using extra hardware resources, so, it is a good step in the right direction